+++ CHANGELOG.md
@@ -15,7 +15,10 @@
         b) User can pass id/offset for app/asset in for array references. For `tx.Accounts` you can pass address directly in teal code.
     * Byteslice arithmetic ops (`b+`, `b-`, `b*`, `b/`, `b%`, `b<`, `b>`, `b<=`, `b>=`, `b==`, `b!=`, `b\`, `b&`, `b^`, `b~`, `bzero`).
     * Additional mathematical opcodes: `divmodw`, `exp`, `expw`, `shl`, `shr`, `sqrt`
+    * More Versatile Global and Local Storage (combination of upto 128 bytes allowed between key-value).
+    * Asset URL change (max size increased till 96 bytes).
 + Updated all examples & tests to use TEALv4 (`#pragma version 4`)
++ TEALv4: gaid, gaids opcodes (knowable creatable id)
 
 
 ### Bug Fixes
@@ -34,6 +37,11 @@
 
 ### Dependencies
 * Upgraded pyTEAL version [`0.8.0`](https://github.com/algorand/pyteal/releases/tag/v0.8.0) in pipfile.
+
+## v1.2.1 2021-09-15
+
+### Bug Fixes
+* Fix `algob init <project-name>`.
 
 ## v1.2.0 2021-08-09